--- scientia/src/components/PassCourse/passCourses.module.css
@@ -1,30 +1,0 @@
-.contenedor {
-    margin-left:15%;
-    margin-right:15%;
-}
-
-.botones {
-    display: flex;
-    justify-content: space-around;
-    margin-bottom: 15px;
-}
-
-.precio {
-    border: 1px solid;
-    display: inline-block;
-    margin-top: 20px;
-    padding: 13px;
-    flex-wrap: nowrap;
-    font-weight: bold;
-}
-
-.contenedorPrecio{
-    display: flex;
-    justify-content: center;
-}
-
-.contenedorModal{
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-}
