--- scientia/src/components/Profiles/AdminPanel/CourseManagement/CourseManagement.css
@@ -1,10 +1,0 @@
-.div-coursesmanagement {
-    width: 100%;
-    background-color: rgb(209, 209, 209);
-    text-align: center;
-    padding: 40px 20px 20px 20px;
-}
-
-p {
-    font-size: 20px;
-}
