+++ scientia/src/components/Profiles/AdminPanel/CourseManagement/CourseManagement.module.css
@@ -1,0 +1,55 @@
+.div-coursesmanagement {
+    width: 100%;
+    background-color: rgb(209, 209, 209);
+    text-align: center;
+    padding: 40px 20px 20px 20px;
+}
+
+p {
+    font-size: 20px;
+}
+
+.contenedor {
+    width: 100%;
+    background-color: rgb(209, 209, 209);
+    text-align: center;
+    padding: 40px 20px 20px 20px;
+    overflow-x: scroll;
+    overflow-y: scroll;
+}
+
+.botones {
+    display: flex;
+    justify-content: space-around;
+    margin-bottom: 15px;
+}
+
+.precio {
+    border: 1px solid;
+    display: inline-block;
+    margin-top: 30px;
+    margin-bottom: 30px;
+    padding: 13px;
+    flex-wrap: nowrap;
+    font-weight: bold;
+}
+
+.contenedorPrecio{
+    display: flex;
+    justify-content: center;
+}
+
+.contenedorModal{
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+}
+
+.tarjetas{
+    display: flex;
+    flex-wrap: nowrap !important;
+}
+
+.tarjeta {
+    width: 600px;
+}
