+++ scientia/src/router/App.js
@@ -20,7 +20,6 @@ import ResponsivePlayer from '../components/Player/Player';
 import UserPanel from '../components/Profiles/UserPanel/UserPanel';
 import AdminPanel from '../components/Profiles/AdminPanel/AdminPanel';
 import Payment from '../components/Payment/payment';
-import PassCourses from '../components/PassCourse/PassCourses';
 
 function App() {
   return (
@@ -78,11 +77,6 @@ function App() {
 
       <Route path='/adminprofile'>
         <AdminPanel />
-      </Route>
-
-            
-      <Route path='/passCourses'>
-        <PassCourses />
       </Route>
       
       <Footer />
